refactor: Deprecates .NET Framework target
This project was already quite out of date from when I first
worked on it. A _lot_ has changed in the .NET ecosystem!

There were a few reasons why I wanted to do this:

	- I wanted to more easily work towards a build pipeline
	- I want less reliance on the GUI for build related tasks
	- .NET Framework seems to be a thing of the past

Initially, the upgrade assistant was used to upgrade the project,
but it wasn't clever enough to know what was required and what
could be discarded. I suppose I had let the project collect too
much dust and warts. This didn't work out, because there were too
many old references that perhaps were no longer required.

The second attempt involved me simply creating and entirely new
solution with `dotnet`, and carrying over 1 by 1 each file,
confirming that I could get a build of both `Debug` and `Release`
configurations, along with passing tests (again, via `dotnet`).
This resulted in a cleaner baseline project.

From there, I copied over the files that were still relevant, and
removed the ones that were not from here.

As part of this, the tests were upgraded to use xunit instead of
MSBuild. This seems to be what all the cool kids are playing with
these days, and the modifications to work with it were minimal.

But of course, the big question: what did I replace .NET Framework
with? Now we build towards 3 targets:

	1. .NET 6.0
	2. .NET Standard 2.0
	3. .NET Standard 2.1

If you can, I suggest using the .NET 6.0 build. If however you are
like me, and primarily use C# with Unity, then as of this writing you
should use .NET Standard 2.1 build.