ci: Add first pipeline
This project should not be in a state where a pipeline can actually be added for it. For now, I am just interested in getting a base pipeline set up using one of the Github Action templates.

This (hopefully) should build and test. I half expect it to not function properly, but I find it is easier to add these to `master` first, and then iterate on it in a separate branch.