ci: Adds SonarCloud analysis
The simple build and test workflow from before is replaced with
a slightly modified version .NET build template that SonarCloud
provides when you set up a project.

It differs slightly than what is suggested, as it also
incorporates coverage report generation.

I expect there to be quality gate issues on this, as this is the
first time the scan will apply to master. Most of this was tested
on a branch first, but the scan results dont really show until
you get that first scan on master.

Once we get the first set of results, I expect to have a followup
branch and PR to correct some of the issues that come up in that
scan